Suppose an array sorted in ascending order is rotated at some pivot unknown to you beforehand.  (i.e., 0 1 2 4 5 6 7 might become 4 5 6 7 0 1 2).  You are given a target value to search. If found in the array return its index, otherwise return -1.  You may assume no duplicate exists in the array.
As this is the rotated sorted array, we need to find the pivot(position around sorted array is rotated, nothing but smallest element in the array) in binary search way by using logic in rotated sorted array, i.e pivot is less than its prior element. We search in left part of pivot if mid element is less than first element of array. if not we search in right part.