Given an array of meeting time intervals consisting of start and end times, find the minimum number of conference rooms required.
Solve it by sorting those intervals by start indices and maintaining a min heap of end values of interval,   and then removing the top of heap when start of an interval is greater than or equal to end of previous interval, adding the end values to the heap while iterating the sorted intervals. size of the priority queue is our answer.