Given an input string (s) and a pattern (p), implement regular expression matching with support for '.' and '*'.
logic for '.' is straightforward, but for '*' (repetition of preceding character), when we encounter it there are two cases 1) there may be zero repetition of characters in pattern when compared with string s. so update dp value with value of dp when previous two characters(preceding character and '*')  removed in pattern (dp[i][j-2]). 2) other case is when corresponding character may be part of repetition. so we compare the preceding character in pattern with current character in string if they are same (or even when preceding character is '.'), we update dp with value of dp when current character of string removed.