Given a binary search tree (BST), find the lowest common ancestor (LCA) of two given nodes in the BST.  According to the definition of LCA on Wikipedia: “The lowest common ancestor is defined between two nodes p and q as the lowest node in T that has both p and q as descendants (where we allow a node to be a descendant of itself).”  Given binary search tree:  root = [6,2,8,0,4,7,9,null,null,3,5]
same logic as LCA for binary tree but here we are at more ease in BST. once we know that p and q exist in left and right subtrees of a node, we can directly return that node. we also check whether p and q lie on the same side and return a node which ever we find first.